+++ examples/src/AnimatedExample/index.js
@@ -1,9 +1,8 @@
 import React, { Component } from "react";
 import { Stage } from "react-pixi-fiber";
-import Animated from "animated";
-import AnimatedSprite from "./AnimatedSprite";
+import Animated from "./animatedPixiTarget";
 import * as PIXI from "pixi.js";
-import bunny from "../Bunny/bunny.png";
+import Bunny from "../Bunny";
 
 const OPTIONS = {
   backgroundColor: 0x1099bb,
@@ -13,35 +12,35 @@ const centerAnchor = new PIXI.Point(0.5, 0.5);
 
 class BunnyExample extends Component {
   state = {
-    rotation: new Animated.Value(0),
+    animationProgress: new Animated.Value(0),
   };
 
   handleDown = () => {
-    Animated.spring(this.state.rotation, { toValue: 1 }).start();
+    Animated.spring(this.state.animationProgress, { toValue: 1 }).start();
   };
 
   handleUp = () => {
-    Animated.spring(this.state.rotation, { toValue: 0 }).start();
+    Animated.spring(this.state.animationProgress, { toValue: 0 }).start();
   };
 
   render() {
     return (
       <Stage width={800} height={600} options={OPTIONS}>
-        <AnimatedSprite
+        <Bunny
           anchor={centerAnchor}
+          as={Animated.Sprite}
           interactive
           pointerdown={this.handleDown}
           pointerup={this.handleUp}
           position="400,300"
-          rotation={this.state.rotation.interpolate({
+          rotation={this.state.animationProgress.interpolate({
             inputRange: [0, 1],
             outputRange: [0, Math.PI * 2],
           })}
-          scale={this.state.rotation.interpolate({
+          scale={this.state.animationProgress.interpolate({
             inputRange: [0, 1],
-            outputRange: [1, 2],
+            outputRange: [1, 8],
           })}
-          texture={PIXI.Texture.fromImage(bunny)}
         />
       </Stage>
     );