+++ index.d.ts
@@ -47,7 +47,7 @@ declare module 'react-pixi-fiber' {
   export interface ContainerProperties extends ChildlessComponent<PIXI.Container> {}
 
   /**
-   * A component wrapper for `PIXI.extras.BitmapText`.
+   * A component wrapper for `PIXI.Container`.
    *
    * see: http://pixijs.download/dev/docs/PIXI.Container.html
    */
@@ -118,7 +118,7 @@ declare module 'react-pixi-fiber' {
   export class Stage extends React.Component<StageProperties> {}
 
   /** Custom React Reconciler render method. */
-  export function render(pixiElement: PIXI.DisplayObject | PIXI.DisplayObject[], stage: PIXI.Container, callback?: Function): void;
+  export function render(pixiElement: React.ReactElement<any> | React.ReactElement<any>[] | PIXI.DisplayObject | PIXI.DisplayObject[], stage: PIXI.Container, callback?: Function): void;
 
   /**
    * Custom component properties.
@@ -151,4 +151,16 @@ declare module 'react-pixi-fiber' {
      */
     type: string
   ): React.ReactType<T>;
+
+  /**
+   * AppContext
+   */
+  export const AppContext: React.Context<PIXI.Application>;
+
+  /**
+   * BatchedUpdates same as ReactDOM
+   */
+  export function unstable_batchedUpdates<A, B>(callback: (a: A, b: B) => any, a: A, b: B): void;
+  export function unstable_batchedUpdates<A>(callback: (a: A) => any, a: A): void;
+  export function unstable_batchedUpdates(callback: () => any): void;
 }